+++ R/plot-anomaly_diagnostics.R
@@ -35,6 +35,12 @@
 #'  If FALSE, returns a static `ggplot2` plot.
 #' @param .trelliscope Returns either a normal plot or a trelliscopejs plot (great for many time series)
 #'  Must have `trelliscopejs` installed.
+#' @param .trelliscope_params Pass parameters to the `trelliscopejs::facet_trelliscope()` function as a `list()`.
+#'  The only parameters that cannot be passed are:
+#'  - `ncol`: use `.facet_ncol`
+#'  - `nrow`: use `.facet_nrow`
+#'  - `scales`: use `facet_scales`
+#'  - `as_plotly`: use `.interactive`
 #'
 #'
 #' @return A `plotly` or `ggplot2` visualization
@@ -151,7 +157,8 @@ plot_anomaly_diagnostics <- function(
     .color_lab = "Anomaly",
 
     .interactive = TRUE,
-    .trelliscope = FALSE
+    .trelliscope = FALSE,
+    .trelliscope_params = list()
 ) {
 
     # Checks
@@ -212,7 +219,8 @@ plot_anomaly_diagnostics.data.frame <- function(
     .color_lab = "Anomaly",
 
     .interactive = TRUE,
-    .trelliscope = FALSE
+    .trelliscope = FALSE,
+    .trelliscope_params = list()
 ) {
 
     # Tidy Eval Setup
@@ -331,14 +339,27 @@ plot_anomaly_diagnostics.data.frame <- function(
 
     } else {
 
-        g <- g +
-            trelliscopejs::facet_trelliscope(
-                facets    = ggplot2::vars(!!! rlang::syms(facet_names)),
+        # g <- g +
+        #     trelliscopejs::facet_trelliscope(
+        #         facets    = ggplot2::vars(!!! rlang::syms(facet_names)),
+        #         ncol      = .facet_ncol,
+        #         nrow      = .facet_nrow,
+        #         scales    = .facet_scales,
+        #         as_plotly = .interactive
+        #     )
+
+        trell <- do.call(trelliscopejs::facet_trelliscope, c(
+            list(
+                facets    = ggplot2::vars(!!! rlang::syms(group_names)),
                 ncol      = .facet_ncol,
                 nrow      = .facet_nrow,
                 scales    = .facet_scales,
                 as_plotly = .interactive
-            )
+            ),
+            .trelliscope_params
+        ))
+
+        g <- g + trell
 
     }
 
@@ -385,7 +406,8 @@ plot_anomaly_diagnostics.grouped_df <- function(
     .color_lab = "Anomaly",
 
     .interactive = TRUE,
-    .trelliscope = FALSE
+    .trelliscope = FALSE,
+    .trelliscope_params = list()
 ) {
 
 
@@ -445,7 +467,8 @@ plot_anomaly_diagnostics.grouped_df <- function(
         .color_lab          = .color_lab,
 
         .interactive        = .interactive,
-        .trelliscope        = .trelliscope
+        .trelliscope        = .trelliscope,
+        .trelliscope_params = .trelliscope_params
     )
 
 }